django_email-main/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
django-rest-swagger==2.2.0
djangorestframework==3.12.4
greenlet==1.1.2
idna==3.3
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
mssql-django==1.0
openapi-codec==1.3.2
Pillow==8.4.0
psycopg2==2.9.1
pymssql==2.2.2
pyodbc==4.0.32
pytz==2021.3
requests==2.26.0
simplejson==3.17.5
SQLAlchemy==1.4.25
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
